Extract browser drag/drop file handling into engine/src/browserFileDragDrop.ts
Moves initializeBrowserFileHandler, handleBrowserMessage,
setupPanelResizeAndDrag, setupBrowserFileHandlers, handleDragOver,
handleDropEvent, extractDroppedFilePaths, handleDroppedFiles, and
handleBrowserFiles out of main.ts, and drops the now-unused fileHandler
imports.
